coderhouse-final-proyecto-Colombo-Leandro/requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
redshift_connector